+++ app/models/caffeinate/campaign.rb
@@ -21,6 +21,7 @@ module Caffeinate
     has_many :mailings, through: :caffeinate_campaign_subscriptions, class_name: 'Caffeinate::CampaignSubscriptions'
 
     scope :active, -> { where(active: true) }
+
     # Poorly-named Campaign class resolver
     def to_dripper
       ::Caffeinate.dripper_collection.resolve(self)
